Migrate ViewLogsRoute to TypeScript
