+++ client/contexts/ServerContext/endpoints.ts
@@ -10,6 +10,7 @@ import type { DnsEndpoints } from './endpoints/v1/dns';
 import type { EmojiCustomEndpoints } from './endpoints/v1/emojiCustom';
 import type { GroupsEndpoints } from './endpoints/v1/groups';
 import type { ImEndpoints } from './endpoints/v1/im';
+import type { MiscEndpoints } from './endpoints/v1/misc';
 import type { OmnichannelEndpoints } from './endpoints/v1/omnichannel';
 import type { RoomsEndpoints } from './endpoints/v1/rooms';
 import type { TeamsEndpoints } from './endpoints/v1/teams';
@@ -29,7 +30,8 @@ type Endpoints = ChatEndpoints &
 	UsersEndpoints &
 	EngagementDashboardEndpoints &
 	AppsEndpoints &
-	OmnichannelEndpoints;
+	OmnichannelEndpoints &
+	MiscEndpoints;
 
 type Endpoint = UnionizeEndpoints<Endpoints>;
