+++ client/contexts/ServerContext/endpoints/v1/misc.ts
@@ -1,0 +1,11 @@
+export type MiscEndpoints = {
+	'stdout.queue': {
+		GET: () => {
+			queue: {
+				id: string;
+				string: string;
+				ts: Date;
+			}[];
+		};
+	};
+};
